Resized map size + Added body backgound
Resized map size from 10000 to 2000 for better playtime, also added
background to body to show map when at edges.
Also removed out of bounds code and added check on movements to only
move if it will be a valid move meaning you just stip before the edge of
the map making cleaner gameplay.